+++ components/molecules/index.ts
@@ -1,1 +1,2 @@
 export * from './InputWithLabel';
+export * from './LinkButton';
